% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogam-package.R
\docType{package}
\name{autogam-package}
\alias{autogam-package}
\title{Automate the Creation of Generalized Additive Models (GAMs)}
\description{
This wrapper package for 'mgcv' makes it easier to create high-performing Generalized Additive Models (GAMs). With its central function autogam(), by entering just a dataset and the name of the outcome column as inputs, 'AutoGAM' tries to automate the procedure of configuring a highly accurate GAM which performs at reasonably high speed, even for large datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tripartio/autogam}
  \item \url{https://tripartio.github.io/autogam/}
  \item Report bugs at \url{https://github.com/tripartio/autogam/issues}
}

}
\author{
Chitu Okoli \email{Chitu.Okoli@skema.edu}
}
\keyword{internal}
