% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGAM_fit.R
\name{autoGAM_fit}
\alias{autoGAM_fit}
\title{autoGAM_fit}
\usage{
autoGAM_fit(object)
}
\arguments{
\item{object}{An object of class \code{autoGAM_frame}.}
}
\value{
Adds the final fitted model to the \code{autoGAM_frame} and prints out the summary of the final fit.
}
\description{
A wrapper function to fit final (best) Generalized Additive Model (GAM) from an object of class \code{autoGAM_frame}.
}
\examples{
\dontrun{

my_mtcars <- mtcars \%>\% mutate_at('vs',as.factor)

carsGAM <- autoGAM_frame(mpg~disp+drat+vs,data=my_mtcars)

autoGAM_fit(carsGAM)

backward_select(carsGAM) \%>\% autoGAM_fit

}

}
\seealso{
\url{https://shahin-roshani.github.io/autoGAM/articles/autoGAM.html}
}
\author{
Shahin Roshani
}
