% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGAM_frame.R
\name{autoGAM_frame}
\alias{autoGAM_frame}
\title{autoGAM_frame}
\usage{
autoGAM_frame(
  formula,
  resp.base = NULL,
  forms = list("identity", "logb", "exp", power = 2:3, poly = 2:3),
  data,
  ignore.outliers = F,
  family = gaussian(link = "identity"),
  metric = "AIC",
  raw.poly = F,
  interval.alpha = 0.05,
  parallel = F,
  core.nums = NULL
)
}
\arguments{
\item{formula}{A simple formula object with one response and predictor(s).}

\item{resp.base}{Base level of the binary response variable. Default is \code{NULL}.}

\item{forms}{A named list of continuous predictor(s) form(s) with their respected degree(s) or degree(s) of freedom. Famous functions that one can think of are \code{identity()}, \code{log()}, \code{log2()}, \code{logb()}, \code{exp()}, \code{bs()} & \code{ns()} (from splines package), \code{s()} (from gam package) and \code{cut()}. Default is \code{list('identity','logb','exp','power'=2:3,'poly'=2:3)}.
The names of the list are the function(s) names and their respected value(s) are their degree(s)/degree(s) of freedom. Only functions with the form: \emph{f(x)} or \emph{f(x,degree)} with a vector or a matrix as their output should be included in this list.
Note: Functions with the \emph{f(x)} form can be passed in 2 ways: 1. \code{'f'=c()} or 2. \code{'f'}. Any function of the form \emph{f(vector,degree)} should be passed as \code{'f'=}desired degree(s).
If a function of the form \emph{f(vector,degree)} is passed as a single character, the default value for degree will be used and if there's no default value for degree, an error will occur.}

\item{data}{Dataset containing the response and all the predictors that were included in the formula.}

\item{ignore.outliers}{Logical indicating whether outliers should be ignored during the evaluation process of predictors forms or not. Default is \code{FALSE}. When \code{TRUE}, outliers in case of response are detected by \code{car::outlierTest()} which is based on studentized residuals of records and outliers in case of the predictor are determined as records that have \emph{hat-values > 2p} (where \emph{p} is the number of parameters inside the model).}

\item{family}{Family for the response variable in model fits. Default is \code{gaussian(link='identity')}.}

\item{metric}{The name of the metric to be used for evaluation of GLMs performances. Valid values are \code{'AIC'}, \code{'BIC'} & \code{'AICc'}. Default is \code{'AIC'}.}

\item{raw.poly}{Logical indicating whether raw forms of polynomials should be included when polynomial forms (\code{'poly'}) are being evaluated. Default is \code{FALSE}.}

\item{interval.alpha}{Numerical value of alpha for the creation of confidence intervals of predictions. Default value is \emph{0.05}.}

\item{parallel}{Logical indicating whether the evaluation process of different forms must be done in parallel mode or not. Default is \code{FALSE}.}

\item{core.nums}{Number of cores to be used in parallelization process. The default is \code{NULL} where automatically half of CPU cores will be used unless user specifies its value.}
}
\value{
A comprehensive list containing information of the whole evaluation process. By default, autoGAM shows final best forms of continuous predictors and the final set of categorical variables but they are part of a bigger list. You can access all items of the list via:

\enumerate{

\item \code{$data}: Dataset that was used to create different models in the evaluation process.

\item \code{$`forms info`}: A nested data frame including full information of evaluation process. It includes values and predictions for all form(s) on all predictor(s).

\item \code{$`best forms`}: Final best form of continuous predictors that were obtained from the evaluation process.

\item \code{$`final predictors`}: Final predictors (best form of continuous predictors and categorical predictors) that are included in the best GAM model. If the backward argument was set to FALSE, best forms of continuous predictors (and possibly the categorical variables) are returned.

\item \code{$`response family`}: The family of the model's response. This item is for internal use!

}
}
\description{
Automatic evaluation of different forms (functions) of predictors to obtain the best Generalized Additive Model (GAM).
}
\examples{
\dontrun{ autoGAM_frame(mpg~disp+drat+vs,data=mtcars \%>\% mutate_at('vs',as.factor)) }

}
\seealso{
\url{https://shahin-roshani.github.io/autoGAM/articles/autoGAM.html}
}
\author{
Shahin Roshani
}
