% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_select.R
\name{backward_select}
\alias{backward_select}
\title{backward_select}
\usage{
backward_select(
  object,
  data = NULL,
  backward.test = "LRT",
  backward.alpha = 0.05,
  family = gaussian(link = "identity")
)
}
\arguments{
\item{object}{An object of class \code{autoGAM_frame} or a formula of the full model to be processed via backward elimination.}

\item{data}{Data needed for creation of models. Default is \code{NULL} for objects of class \code{autoGAM_frame} where data will be automatically retrieved from the object but data needs to be specified when the \code{object} argument is a formula.}

\item{backward.test}{Test to obtain P-values of predictors during the backward elimination process. Valid values are \code{'LRT'}, \code{'Rao'}, \code{'Chisq'} & \code{'F'}. Default is \code{'LRT'}. See \code{?drop1} for more details.}

\item{backward.alpha}{The maximum alpha value that predictors with P-values greater than that will be sequentially removed. Default value is \emph{0.05}.}

\item{family}{The family of the response variable. Default is \code{gaussian(link='identity')} for situations where \code{object} argument is a formula (for formulas with responses from other classes, you need to specify this argument). When \code{object} is from class \code{autoGAM_frame}, default value will be automatically replaced by the true family coming from the object.}
}
\value{
If the object is from class \code{autoGAM_frame}, the \code{$`best forms`} & \code{$`final predictors`} parts of the object will be updated and a list of sequential results of backward elimination process will be added to the \code{autoGAM_frame} list. If the object is a formula, the mentioned sequential list of backward process will be returned.
}
\description{
P-value based backward selection on Generalized Linear Models (GLMs).
}
\examples{
\dontrun{

my_mtcars <- mtcars \%>\% mutate_at('vs',as.factor)

carsGAM <- autoGAM_frame(mpg~disp+drat+vs,data=my_mtcars)

backward_select(carsGAM)

#Or for example:

backward_select(mpg~poly(disp,2)+I(drat^3)+vs,data=my_mtcars)

my_iris <- iris \%>\% filter(Species!='setosa')

backward_select(Species~ns(Sepal.Length,3)+cut(Petal.Width,2),data=my_iris,family=binomial(link='logit'))

}

}
\seealso{
\url{https://shahin-roshani.github.io/autoGAM/articles/autoGAM.html}
}
\author{
Shahin Roshani
}
