% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogam.R
\name{autogam generic methods}
\alias{autogam generic methods}
\alias{anova.autogam}
\alias{coef.autogam}
\alias{cooks.distance.autogam}
\alias{formula.autogam}
\alias{influence.autogam}
\alias{logLik.autogam}
\alias{model.matrix.autogam}
\alias{predict.autogam}
\alias{residuals.autogam}
\alias{vcov.autogam}
\title{Generic autogam methods passed on to mgcv::gam methods}
\usage{
\method{anova}{autogam}(object, ...)

\method{coef}{autogam}(object, ...)

\method{cooks.distance}{autogam}(model, ...)

\method{formula}{autogam}(x, ...)

\method{influence}{autogam}(model, ...)

\method{logLik}{autogam}(object, ...)

\method{model.matrix}{autogam}(object, ...)

\method{predict}{autogam}(object, ...)

\method{residuals}{autogam}(object, ...)

\method{vcov}{autogam}(object, ...)
}
\arguments{
\item{object, model}{An object of class \code{autogam}.}

\item{...}{other arguments}

\item{x}{formula}
}
\value{
Returns the return object of the corresponding \code{mgcv::gam} method.
}
\description{
An \code{autogam} object contains a \code{gam} element that is simply an \code{mgcv::gam} object. So, it supports all \code{mgcv::gam} methods by, in most cases, simply passing the \code{gam} element on to their corresponding \code{mgcv::gam} methods. Only the following methods have special specifications for autogam (see their dedicated documentation files for details): \code{\link[=print.autogam]{print.autogam()}}.
}
