% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.autoGAM_frame.R
\name{plot.autoGAM_frame}
\alias{plot.autoGAM_frame}
\title{plot.autoGAM_frame}
\usage{
\method{plot}{autoGAM_frame}(object, type = "final")
}
\arguments{
\item{object}{An object of class \code{autoGAM_frame}.}

\item{type}{The type of the plot to be created. Valid inputs are \code{'forms'} that returns the whole plots of evaluated continuous predictors & \code{'final'} that returns the plots related to the best forms obtained from the evaluation process.}
}
\value{
Desired plots that were requested by the \code{type} argument.
}
\description{
Plot method for an object of class \code{autoGAM_frame}.
}
\examples{
\dontrun{

my_mtcars <- mtcars \%>\% mutate_at('vs',as.factor)

carsGAM <- autoGAM_frame(mpg~disp+drat+vs,data=my_mtcars) \%>\% backward_select \%>\% autoGAM_fit

plot(carsGAM,type='forms')

plot(carsGAM)

}

}
\seealso{
\url{https://shahin-roshani.github.io/autoGAM/articles/autoGAM.html}
}
\author{
Shahin Roshani
}
