% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogam.R
\name{print.autogam}
\alias{print.autogam}
\title{Print Method for autogam Objects}
\usage{
\method{print}{autogam}(x, ...)
}
\arguments{
\item{x}{An object of class \code{autogam}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
This function prints an \code{autogam} object. It calls the \code{mgcv::gam} object \code{print()} method and then adds basic performance metrics from the \code{autogam} object:
\itemize{
\item For models that predict numeric outcomes, it prints "MAE", the mean absolute error, and "Std. accuracy", the standardized accuracy (staccuracy) of the winsorized MAE relative to the mean absolute deviation.
\item For models that predict binary outcomes, it prints "AUC", the area under the ROC curve.
}
}
