% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogam.R
\name{summary.autogam}
\alias{summary.autogam}
\title{Summary Method for autogam Objects}
\usage{
\method{summary}{autogam}(object, ...)
}
\arguments{
\item{object}{An object of class \code{autogam}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
Same return object as \code{\link[mgcv:summary.gam]{mgcv::summary.gam()}}.
}
\description{
This function returns a summary of an \code{autogam} object. It calls the \code{mgcv::gam} object \code{\link[mgcv:summary.gam]{mgcv::summary.gam()}} method.
}
