% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGAM.R
\name{autoGAM}
\alias{autoGAM}
\title{autoGAM}
\usage{
autoGAM(
  resp,
  resp.base = NULL,
  cont.vars = c(),
  cat.vars = c(),
  forms = list("identity", "logb", "exp", power = 2:3, poly = 2:3),
  data,
  ignore.outliers = F,
  family = gaussian(link = "identity"),
  raw.poly = F,
  forms.plot = F,
  backward = T,
  backward.test = "LRT",
  backward.alpha = 0.05,
  interval.alpha = 0.05
)
}
\arguments{
\item{resp}{Name of the response variable.}

\item{resp.base}{Base level of the binary response variable, default is NULL.}

\item{cont.vars}{Vector containing names of continuous variables, default is c(). Also see data argument.}

\item{cat.vars}{Vector containing names of categorical variables, default is c(). Also see data argument.}

\item{forms}{A named list of continuous predictor(s) form(s) with their respected degree(s) or degree(s) of freedom. Famous functions that one can think of are identity, log, log2, logb, exp, bs & ns (from splines package), s (from gam package) and cut,default is list('identity','logb','exp','power'=2:3,'poly'=2:3).
The names of the list are the function(s) names and their respected value(s) are their degree(s)/degree(s) of freedom. Only functions with the form: f(x) or f(x,degree) with a vector or a matrix as their output should be included in this list.
Note: Functions with the f(x) form can be passed in 2 ways: 1. 'f'=c() or 2. 'f'. Any function of the form f(vector,degree) should be passed as 'f'=desired degree(s).
If a function of the form f(vector,degree) is passed as a single character, the default value for degree will be used and if there's no default value for degree, an error will occur.}

\item{data}{Dataset containing the response and all the predictors.
Note: When the cont.vars & cat.vars arguments are BOTH left empty, variables and their respected types that are given in the dataset will be automatically used to distinguish continuous and categorical variables that are available in the whole dataset.
If only one of two cont.vars & cat.vars was left empty, remaining variables in the dataset other than the filled argument (either cont.vars or cat.vars) and the response will be discarded. It is recommended to let autoGAM decide only when your data object only consists of your desired response & predictors with right types.}

\item{ignore.outliers}{Logical indicating whether outliers should be ignored during the evaluation process of predictors forms or not, default is FALSE. When TRUE, outliers in case of response are detected by car::outlierTest which is based on studentized residuals of records and outliers in case of the predictor are determined as records that have hat-values > 2p (where p is the number of parameters inside the model).}

\item{family}{Family for the response variable in GLM fits, default is gaussian(link='identity').}

\item{raw.poly}{Logical indicating whether raw forms of polynomials should be included when polynomial forms ('poly') are being evaluated, default is FALSE.}

\item{forms.plot}{Logical indicating whether plots for all evaluated forms on all continuous predictors should be included in the output or not, default is FALSE.}

\item{backward}{Logical indicating whether backward elimination should be done on the obtained best forms of predictors, default is TRUE.}

\item{backward.test}{Name of the test to be used in the backward elimination process on GLM fits. Valid inputs are 'LRT', 'Rao', 'Chisq' and 'F', default is 'LRT'.}

\item{backward.alpha}{Numerical value of alpha for the backward elimination process, default is 0.05.}

\item{interval.alpha}{Numerical value of alpha for the creation of confidence intervals of predictions.}
}
\value{
A comprehensive list containing information of the whole evaluation process, backward elimination process and the final GAM model. Result includes:

1- $`Forms data`: A nested data frame including full information of evaluation process. It includes values and predictions for all form(s) on all predictor(s).

2- $`Forms plot(s)`: Plots of the dataset in $`Forms data` (Included if argument: forms.plot=TRUE).

3- $`Backward info`: Step by step list of information regarding the backward elimination process (elimination alpha = backward.alpha) (Included if argument: backward=TRUE).

4- $`Final predictors`: Final predictors (best form of continuous predictors and categorical predictors) that are included in the best GAM model. If the backward argument was set to FALSE, best forms of continuous predictors (and possibly the categorical variables) are returned.

5- $`Best forms plot(s)`: Plots with corresponding confidence intervals (alpha = interval.alpha) of predictions for final best model of each continuous predictor.

6- $`Final GAM`: Final fitted GAM based on final predictors in the $`Final predictors` part.
}
\description{
Automatic evaluation of different forms (functions) of predictors to obtain the best Generalized Additive Model (GAM).
}
\examples{
autoGAM(resp='mpg',cont.vars=names(mtcars)[3:6],cat.vars=names(mtcars)[8],
data=mtcars,forms.plot=TRUE)

}
\author{
Shahin Roshani
}
